Refactor generators and add origin for Gazebo 8 RNDF plugin
