+++ terminus/generators/sdf_generator_gazebo_8.py
@@ -1,29 +1,15 @@
-from file_generator import FileGenerator
+from abstract_sdf_generator import AbstractSDFGenerator
 
 
-class SDFGeneratorGazebo8(FileGenerator):
+class SDFGeneratorGazebo8(AbstractSDFGenerator):
     """This class generates an sdf file that loads the road network from an RNDF
     file. In order to use this you will need Gazebo 8 and the plugin to render
     RNDF"""
 
-    def __init__(self, city, rndf_path):
+    def __init__(self, city, rndf_origin, rndf_path):
         super(SDFGeneratorGazebo8, self).__init__(city)
         self.rndf_path = rndf_path
-
-    def end_document(self):
-        self._wrap_document_with_template('document')
-
-    def end_city(self, city):
-        self._wrap_document_with_contents_for(city)
-
-    def end_block(self, block):
-        self._append_to_document(self._contents_for(block))
-
-    def end_building(self, building):
-        self._append_to_document(self._contents_for(building))
-
-    def end_ground_plane(self, ground_plane):
-        self._append_to_document(self._contents_for(ground_plane))
+        self.rndf_origin = rndf_origin
 
     def city_template(self):
         return """
@@ -44,154 +30,8 @@ class SDFGeneratorGazebo8(FileGenerator):
           <waypoints>true</waypoints>
           <junctions>true</junctions>
           <waypoints_material>Gazebo/White</waypoints_material>
+          <origin>{{generator.rndf_origin[0]}} {{generator.rndf_origin[1]}} 0.0</origin>
         </plugin>
           {{inner_contents}}
         </world>
         """
-
-    def block_template(self):
-        return """
-        <model name="{{model.name}}">
-          <static>1</static>
-          <allow_auto_disable>1</allow_auto_disable>
-          <link name="{{model.name}}_link">
-            <pose frame="">
-              {{model.origin.x}} {{model.origin.y}} {{model.origin.z}}
-              0 0 0
-            </pose>
-            <collision name="{{model.name}}_collision">
-              <geometry>
-                <polyline>
-                {% for vertex in model.vertices %}
-                  <point>{{vertex.x}} {{vertex.y}}</point>
-                {% endfor %}
-                <height>{{model.height}}</height>
-                </polyline>
-              </geometry>
-            </collision>
-            <visual name="{{model.name}}_visual">
-              <material>
-                <script>
-                  <uri>
-                    file://media/materials/scripts/gazebo.material
-                  </uri>
-                  <name>Gazebo/Grey</name>
-                </script>
-                <ambient>1 1 1 1</ambient>
-              </material>
-              <meta>
-                <layer>0</layer>
-              </meta>
-              <geometry>
-                <polyline>
-                {% for vertex in model.vertices %}
-                  <point>{{vertex.x}} {{vertex.y}}</point>
-                {% endfor %}
-                <height>{{model.height}}</height>
-                </polyline>
-              </geometry>
-            </visual>
-          </link>
-        </model>"""
-
-    def building_template(self):
-        return """
-          <model name="{{model.name}}">
-             <static>1</static>
-             <link name="{{model.name}}_link">
-                <pose frame="">
-                  {{model.origin.x}} {{model.origin.y}} {{model.origin.z}}
-                  0 0 0
-                </pose>
-                <collision name="{{model.name}}_collision">
-                   <geometry>
-                        <polyline>
-                        {% for vertex in model.vertices %}
-                          <point>{{vertex.x}} {{vertex.y}}</point>
-                        {% endfor %}
-                        <height>{{model.height}}</height>
-                        </polyline>
-                   </geometry>
-                </collision>
-                <visual name="{{model.name}}_visual">
-                   <material>
-                      <script>
-                        <uri>
-                          file://media/materials/scripts/gazebo.material
-                        </uri>
-                        <name>Gazebo/Grey</name>
-                      </script>
-                      <ambient>1 1 1 1</ambient>
-                   </material>
-                   <meta>
-                      <layer>0</layer>
-                   </meta>
-                   <geometry>
-                        <polyline>
-                        {% for vertex in model.vertices %}
-                          <point>{{vertex.x}} {{vertex.y}}</point>
-                        {% endfor %}
-                        <height>{{model.height}}</height>
-                        </polyline>
-                   </geometry>
-                </visual>
-             </link>
-          </model>"""
-
-    def ground_plane_template(self):
-        return """
-          <model name="{{model.name}}">
-            <static>1</static>
-            <pose frame=''>{{model.origin.x}} {{model.origin.y}} 0 0 0 0</pose>
-            <link name='{{model.name}}_link'>
-              <collision name='{{model.name}}'>
-                <geometry>
-                  <plane>
-                    <normal>0 0 1</normal>
-                    <size>{{model.size}} {{model.size}}</size>
-                  </plane>
-                </geometry>
-                <surface>
-                  <friction>
-                    <ode>
-                      <mu>100</mu>
-                      <mu2>50</mu2>
-                    </ode>
-                    <torsional>
-                      <ode/>
-                    </torsional>
-                  </friction>
-                  <contact>
-                    <ode/>
-                  </contact>
-                  <bounce/>
-                </surface>
-                <max_contacts>10</max_contacts>
-              </collision>
-              <visual name='{{model.name}}'>
-                <cast_shadows>0</cast_shadows>
-                <geometry>
-                  <plane>
-                    <normal>0 0 1</normal>
-                    <size>{{model.size}} {{model.size}}</size>
-                  </plane>
-                </geometry>
-                <material>
-                  <script>
-                    <uri>file://media/materials/scripts/gazebo.material</uri>
-                    <name>Gazebo/Grey</name>
-                  </script>
-                </material>
-                <transparency>1</transparency>
-              </visual>
-              <self_collide>0</self_collide>
-              <kinematic>0</kinematic>
-            </link>
-          </model>"""
-
-    def document_template(self):
-        return """
-        <?xml version="1.0" ?>
-        <sdf version="1.5">
-          {{inner_contents}}
-        </sdf>"""
